Improved tests plot + Changed statistical test to Wilcoxon rank-sum test for a more rigorous test + Fixed issue with float-double precision (actually the java version is a mix of float and double) + Added some tests that helped finding float precision issues

+++ src/mofgbmlpy/data/input.py
@@ -44,8 +44,8 @@ class Input:
             pattern_id = 0
             patterns = []
             for row in reader:
-                in_vector = np.zeros(num_dim, dtype=np.float32)
-                c_vector = np.zeros(num_classes, dtype=np.int_)
+                in_vector = np.zeros(num_dim, dtype=np.float64)
+                c_vector = np.zeros(num_classes, dtype=np.int32)
 
                 for i in range(len(in_vector)):
                     in_vector[i] = float(row[i])
@@ -86,7 +86,7 @@ class Input:
             patterns = []
 
             for row in reader:
-                in_vector = np.zeros(num_dim, dtype=np.float32)
+                in_vector = np.zeros(num_dim, dtype=np.float64)
 
                 for i in range(len(in_vector)):
                     in_vector[i] = float(row[i])
