+++ src/mofgbmlpy/fuzzy/rule/rule_basic.pxd
@@ -5,5 +5,5 @@ import cython
 
 
 cdef class RuleBasic(AbstractRule):
-    cpdef float get_fitness_value(self, float[:] attribute_vector)
+    cpdef double get_fitness_value(self, double[:] attribute_vector)
