+++ src/mofgbmlpy/fuzzy/rule/rule_multi.pxd
@@ -4,4 +4,4 @@ cimport numpy as cnp
 import cython
 
 cdef class RuleMulti(AbstractRule):
-    cpdef float get_fitness_value(self, float[:] attribute_vector)
+    cpdef double get_fitness_value(self, double[:] attribute_vector)
