+++ src/mofgbmlpy/gbml/objectives/pittsburgh/num_rules.pxd
@@ -2,4 +2,4 @@ from mofgbmlpy.gbml.objectives.objective_function cimport ObjectiveFunction
 from mofgbmlpy.gbml.solution.abstract_solution cimport AbstractSolution
 
 cdef class NumRules(ObjectiveFunction):
-    cpdef void run(self, AbstractSolution[:] solutions, int obj_index, float[:] out)
+    cpdef void run(self, AbstractSolution[:] solutions, int obj_index, double[:] out)
