+++ src/mofgbmlpy/gbml/problem/pittsburgh_problem.pyx
@@ -118,11 +118,11 @@ class PittsburghProblem(Problem):
 
         Args:
             X (Population): Population evaluated
-            out (float[,]): Objective function values for each solution
+            out (double[,]): Objective function values for each solution
             *args (tuple): Other arguments for Pymoo
             **kwargs (dict): Other arguments for Pymoo
         """
-        cdef cnp.ndarray[float, ndim=2] eval_values = np.empty((len(X), self.get_num_objectives()), dtype=np.float32)
+        cdef cnp.ndarray[double, ndim=2] eval_values = np.empty((len(X), self.get_num_objectives()), dtype=np.float64)
         cdef int i
 
         # Update eval values
